Suppressed some Windows warnings.

+++ src/service/usServiceReference.cpp
@@ -176,8 +176,8 @@ std::ostream& operator<<(std::ostream& os, const ServiceReference& serviceRef)
      << " (";
   std::vector<std::string> keys;
   serviceRef.GetPropertyKeys(keys);
-  int keySize = keys.size();
-  for(int i = 0; i < keySize; ++i)
+  size_t keySize = keys.size();
+  for(size_t i = 0; i < keySize; ++i)
   {
     os << keys[i] << "=" << serviceRef.GetProperty(keys[i]);
     if (i < keySize-1) os << ",";
